Performs join/merge for disk.frames

Performs join/merge for disk.frames

# S3 method for disk.frame
anti_join(
  x,
  y,
  by = NULL,
  copy = FALSE,
  ...,
  outdir = tempfile("tmp_disk_frame_anti_join"),
  merge_by_chunk_id = FALSE,
  overwrite = TRUE,
  .progress = FALSE
)

# S3 method for disk.frame
full_join(
  x,
  y,
  by = NULL,
  copy = FALSE,
  ...,
  outdir = tempfile("tmp_disk_frame_full_join"),
  overwrite = TRUE,
  merge_by_chunk_id,
  .progress = FALSE
)

# S3 method for disk.frame
inner_join(
  x,
  y,
  by = NULL,
  copy = FALSE,
  suffix = c(".x", ".y"),
  ...,
  keep = FALSE,
  outdir = tempfile("tmp_disk_frame_inner_join"),
  merge_by_chunk_id = NULL,
  overwrite = TRUE,
  .progress = FALSE
)

# S3 method for disk.frame
left_join(
  x,
  y,
  by = NULL,
  copy = FALSE,
  suffix = c(".x", ".y"),
  ...,
  keep = FALSE,
  outdir = tempfile("tmp_disk_frame_left_join"),
  merge_by_chunk_id = FALSE,
  overwrite = TRUE,
  .progress = FALSE
)

# S3 method for disk.frame
semi_join(
  x,
  y,
  by = NULL,
  copy = FALSE,
  ...,
  outdir = tempfile("tmp_disk_frame_semi_join"),
  merge_by_chunk_id = FALSE,
  overwrite = TRUE,
  .progress = FALSE
)

Arguments

x

a disk.frame

y

a data.frame or disk.frame. If data.frame then returns lazily; if disk.frame it performs the join eagerly and return a disk.frame

by

join by

copy

same as dplyr::anti_join

...

same as dplyr's joins

outdir

output directory for disk.frame

merge_by_chunk_id

the merge is performed by chunk id

overwrite

overwrite output directory

.progress

Show progress or not. Defaults to FALSE

suffix

see dplyr::XXX_join

keep

see dplyr::XXX_join

Value

disk.frame or data.frame/data.table

Examples

df.df = as.disk.frame(data.frame(x = 1:3, y = 4:6), overwrite = TRUE)
df2.df = as.disk.frame(data.frame(x = 1:2, z = 10:11), overwrite = TRUE)

anti_joined.df = anti_join(df.df, df2.df) 
#> Warning: merge_by_chunk_id = FALSE. This will take significantly longer and the preparations needed are performed eagerly which may lead to poor performance. Consider making y a data.frame or set merge_by_chunk_id = TRUE for better performance.
#> Appending disk.frames: 
#> Appending disk.frames: 
#> Error in anti_join(.x, .y, by = by, copy = copy, ..., overwrite = overwrite): `...` must be empty.
#> ✖ Problematic arguments:
#> • ..1 = xch
#> • ..2 = ych
#> • overwrite = overwrite
#> ℹ Did you forget to name an argument?

anti_joined.df %>% collect
#> Error in collect(.): object 'anti_joined.df' not found

anti_joined.data.frame = anti_join(df.df, data.frame(x = 1:2, z = 10:11))

# clean up
delete(df.df)
delete(df2.df)
delete(anti_joined.df)
#> Error in "disk.frame" %in% class(df): object 'anti_joined.df' not found
cars.df = as.disk.frame(cars)

join.df = full_join(cars.df, cars.df, merge_by_chunk_id = TRUE)

# clean up cars.df
delete(cars.df)
delete(join.df)
cars.df = as.disk.frame(cars)

join.df = inner_join(cars.df, cars.df, merge_by_chunk_id = TRUE)
#> Error in inner_join(.x, .y, by = by, copy = copy, suffix = suffix, ...,     keep = keep): `...` must be empty.
#> ✖ Problematic arguments:
#> • ..1 = xch
#> • ..2 = ych
#> ℹ Did you forget to name an argument?

# clean up cars.df
delete(cars.df)
delete(join.df)
#> Error: [ENOENT] Failed to search directory 'C:/Users/RTX2080/AppData/Local/Temp/Rtmp2LOQEk/tmp_disk_frame_full_join361c47bd60a6': no such file or directory
cars.df = as.disk.frame(cars)

join.df = left_join(cars.df, cars.df)
#> Warning: `merge_by_chunk_id = FALSE`. This will take significantly longer and the preparations needed are performed eagerly which may lead to poor performance. Consider making `y` a data.frame or set merge_by_chunk_id = TRUE for better performance.
#> Appending disk.frames: 
#> Appending disk.frames: 
#> Error in left_join(.x, .y, by = by, copy = copy, suffix = suffix, ...,     keep = keep): `...` must be empty.
#> ✖ Problematic arguments:
#> • ..1 = xch
#> • ..2 = ych
#> ℹ Did you forget to name an argument?

# clean up cars.df
delete(cars.df)
delete(join.df)
#> Error: [ENOENT] Failed to search directory 'C:/Users/RTX2080/AppData/Local/Temp/Rtmp2LOQEk/tmp_disk_frame_full_join361c47bd60a6': no such file or directory
cars.df = as.disk.frame(cars)

join.df = semi_join(cars.df, cars.df)
#> Warning: merge_by_chunk_id = FALSE. This will take significantly longer and the preparations needed are performed eagerly which may lead to poor performance. Consider making y a data.frame or set merge_by_chunk_id = TRUE for better performance.
#> Appending disk.frames: 
#> Appending disk.frames: 
#> Error in semi_join(.x, .y, by = by, copy = copy, ..., overwrite = overwrite): `...` must be empty.
#> ✖ Problematic arguments:
#> • ..1 = xch
#> • ..2 = ych
#> • overwrite = overwrite
#> ℹ Did you forget to name an argument?

# clean up cars.df
delete(cars.df)
delete(join.df)
#> Error: [ENOENT] Failed to search directory 'C:/Users/RTX2080/AppData/Local/Temp/Rtmp2LOQEk/tmp_disk_frame_full_join361c47bd60a6': no such file or directory