rbindlist disk.frames together

rbindlist disk.frames together

rbindlist.disk.frame(df_list, outdir = tempfile(fileext = ".df"),
  by_chunk_id = TRUE, parallel = TRUE, compress = 50,
  overwrite = TRUE, .progress = TRUE)

Arguments

df_list
A list of disk.frames

outdir
Output directory of the row-bound disk.frames

by_chunk_id
If TRUE then only the chunks with the same chunk IDs will be bound

parallel
if TRUE then bind multiple disk.frame simultaneously, Defaults to TRUE

compress
0-100, 100 being the highest compression rate.

overwrite
overwrite the output directory

.progress
A logical, for whether or not to print a progress bar for multiprocess, multisession, and multicore plans. From furrr

Examples

cars.df = as.disk.frame(cars)

# row-bind two disk.frames
cars2.df = rbindlist.disk.frame(list(cars.df, cars.df))

#> Appending disk.frames: 

# clean up cars.df
delete(cars.df)
delete(cars2.df)